Major Updates
- Added Levels 1-3
- Added Story for levels 1-3
- Added new music for story and levels
- Created bosses for levels 1-3
- Imported new assets for level 3 parallax background
- Fixed a missing script issue
- Updated victory menu manager to accomodate scene transitions
- Updated build settings for levels 1-3, story 1-3, main menu, and ready
scene
- Updated GameManager to track current level

+++ Assets/Scripts/Managers/Parallax.cs
@@ -10,8 +10,8 @@ public class Parallax : MonoBehaviour {
 	List<GameObject> backgrounds = new List<GameObject>();
 	public Vector3 originalPosition;
 	public float spriteWidth;
-	public int minSpeed;
-	public int maxSpeed;
+	public float minSpeed;
+	public float maxSpeed;
 
 	void Awake(){
 		i = this;
